♻️ remove CustomQuestion type
