+++ src/components/QuestionsForm/Question.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
   import { reverseArray } from '../../utils'
-  import type { CustomQuestion } from '../../pages/UserInfos/questions/ExperienceGrade.svelte'
+  import type { Question } from './QuestionsForm.svelte'
 
-  export let question: CustomQuestion
+  export let question: Question
 
   const HIGHER_GRADE = 6
   const GRADES_RANGE = Array.from(Array(HIGHER_GRADE).keys()) // generate [0,1,2, ..., HIGHER_GRADE-1]
