+++ src/components/QuestionsForm/QuestionsForm.svelte
@@ -4,16 +4,16 @@ import type { Qualification } from '../../const'
 /* eslint-disable no-multiple-empty-lines */
 export interface Question {
     ids: [Qualification, Qualification]
+    labels: [string, string]
     // the labels are inverted
     inverted?: boolean
     grade: number | null
   }
 </script>
 <script lang="ts">
-  import type { CustomQuestion } from '../../pages/UserInfos/questions/ExperienceGrade.svelte'
   import QuestionComponent from './Question.svelte'
 
-  export let questions: CustomQuestion[]
+  export let questions: Question[]
 </script>
 
 <table>
