+++ src/pages/UserInfos/questions/AnyExperience.svelte
@@ -2,11 +2,11 @@
   import { _ } from 'svelte-i18n'
 
   import QuestionsForm from '../../../components/QuestionsForm/QuestionsForm.svelte'
-  import type { CustomQuestion } from '../questions/ExperienceGrade.svelte'
+  import type { Question } from '../../../components/QuestionsForm/QuestionsForm.svelte'
 
   export let grade: number | null = null
 
-  let questions: CustomQuestion[] = [
+  let questions: Question[] = [
     {
       ids: ['any_experience', 'lot_of_experience'],
       labels: [$_('user_infos.any_experience.never'), $_('user_infos.any_experience.often')],
