+++ src/pages/UserInfos/questions/ExperienceGrade.svelte
@@ -1,15 +1,9 @@
-<script context="module" lang="ts">
-  /* eslint-disable no-multiple-empty-lines */
-  import type { Question } from '../../../components/QuestionsForm/QuestionsForm.svelte'
-  
-  export type CustomQuestion = Question & { labels: [string, string] };
-</script>
 <script lang="ts">
   import { _ } from 'svelte-i18n'
-
   import QuestionsForm from '../../../components/QuestionsForm/QuestionsForm.svelte'
+  import type { Question } from '../../../components/QuestionsForm/QuestionsForm.svelte'
 
-  const QUESTIONS: Question[] = [
+  const QUESTIONS: Omit<Question, 'labels'>[] = [
     {
       ids: ['unpleasant', 'pleasant'],
       grade: null,
@@ -42,14 +36,14 @@
     }
   ]
 
-  let questionForForm: CustomQuestion[] = QUESTIONS.map(question => {
+  let questionForForm: Question[] = QUESTIONS.map(question => {
     return {
       ...question,
       labels: [$_(`questions.${question.ids[0]}`), $_(`questions.${question.ids[1]}`)]
     }
   })
 
-  export let questions: Question[]
+  export let questions: Omit<Question, 'labels'>[]
   $: {
     questions = questionForForm.map(q => {
       // eslint-disable-next-line
